Normalize composed decisions from owner packets
